+++ client/components/Static/AuthForm.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { auth } from '../store';
+import { auth } from '../../store';
 
 const AuthForm = props => {
   const { name, displayName, handleSubmit, error } = props;
