+++ client/components/Static/Navbar.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { logout } from '../store';
+import { logout } from '../../store';
 
 class Navbar extends Component {
   handleLogout = () => {
