+++ client/components/UserProfile/UserProfileEdit.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import axios from 'axios';
 import { connect } from 'react-redux';
-import { storage } from '../config/firebase';
+import { storage } from '../../config/firebase';
 
 class UserProfileEdit extends Component {
   constructor(props) {
