+++ client/components/index.js
@@ -1,13 +1,12 @@
-export { default as Navbar } from './Navbar';
-export { default as HomePage } from './HomePage';
-export { default as Templates } from './Templates';
-export { default as UserProfile } from './UserProfile';
+export { default as Navbar } from './Static/Navbar';
+export { Login, Signup } from './Static/AuthForm';
+export { default as UserProfile } from './UserProfile/UserProfile';
+export { default as UserProfileEdit } from './UserProfile/UserProfileEdit';
+export { default as Templates } from './Templates/Templates';
+export { default as SelectTemplate } from './Templates/SelectTemplate';
+export { default as SingleStoryContainer } from './Stories/SingleStory';
+export { default as OpenStories } from './Stories/OpenStories';
+export { default as CompletedChapter } from './Stories/CompletedChapter';
+export { default as HomePage } from './HomePage/HomePage';
 export { default as RootCanvas } from './canvas/RootCanvas';
-export { default as SingleStoryContainer } from './SingleStory';
-export { default as UnitStory } from './UnitStory';
 //export { default as Custom } from './CustomHomeOne';
-export { Login, Signup } from './AuthForm';
-export { default as SelectTemplate } from './SelectTemplate';
-export { default as OpenStories } from './OpenStories';
-export { default as CompletedChapter } from './CompletedChapter';
-export { default as UserProfileEdit } from './UserProfileEdit';
