+++ client/index.js
@@ -5,8 +5,8 @@ import { Router } from 'react-router-dom';
 import history from './history';
 import store from './store';
 import App from './App';
+// import style from './style.css'
 
-// establishes socket connection
 import './socket';
 
 ReactDOM.render(
