+++ package.json
@@ -75,6 +75,7 @@
     "babel-eslint": "^8.2.2",
     "babel-loader": "^8.0.0-beta.2",
     "chai": "^3.5.0",
+    "css-loader": "^0.28.11",
     "enzyme": "^3.0.0",
     "enzyme-adapter-react-16": "^1.0.0",
     "eslint": "^4.19.1",
@@ -89,6 +90,7 @@
     "react-test-renderer": "^16.0.0",
     "redux-devtools-extension": "^2.13.2",
     "redux-mock-store": "^1.2.3",
+    "style-loader": "^0.21.0",
     "supertest": "^3.0.0",
     "webpack": "^4.12.0",
     "webpack-cli": "^3.0.3",
